Move _collect_by_type.
